labelling/requirements.txt
ultralytics>=8.3.0
rasterio>=1.3.0
numpy>=1.24.0

opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0

pyyaml>=6.0
geopandas
sahi

